requirements.txt
Flask==3.1.3
flask-mysqldb==2.0.0